feat: switch to the latest version of the InPost API
BREAKING CHANGE: This updates changes the interface of the library,
dropping the now-redundant `getAvailabilityForLocation` function
and changing the shape of objects returned by existing functions.

This change gets the library working again with the current version
of the InPost private API, as used on InPost's website.

As part of these changes:

- The `getAvailabilityForLocation` function is removed, as other
  functions now return locker availability without an extra call.
- The way that locker availability is described is changed. We now
  expose simple keys for availability (e.g.
  `smallLockerAvailability`) and availability is presented through
  the `LockerAvailabilityLevel` enum, rather than a specific number
  of free lockers.